New Arrow Animation
A new class was made for a second type of animation for the arrow. The new class inherits from the previous arrow animation's class. It is only applied in the phone version. This is for when scanning a physical surface so the player can start the game. The animation, however, is not tested yet.