Toggling Quit Button in Games
In the Base Game Script, I made two functions (one to set and the other to toggle) to make the quit button in the game canvas active or inactive. The first function of them (to set) is used when a game is complete. Thus when a game is complete, the quit button is set to inactive so that the player is sure to be rewarded.